move config, model to files
