+++ src/config.rs
@@ -1,0 +1,58 @@
+#[derive(Debug, serde::Deserialize)]
+pub struct SttConfig {
+    pub audio_silence_prefix_seconds: f64,
+    pub audio_delay_seconds: f64,
+}
+
+#[derive(Debug, serde::Deserialize)]
+pub struct Config {
+    pub card: usize,
+    pub text_card: usize,
+    pub dim: usize,
+    pub n_q: usize,
+    pub context: usize,
+    pub max_period: f64,
+    pub num_heads: usize,
+    pub num_layers: usize,
+    pub causal: bool,
+    pub stt_config: SttConfig,
+}
+
+impl Config {
+    pub fn model_config(&self) -> moshi::lm::Config {
+        let lm_cfg = moshi::transformer::Config {
+            d_model: self.dim,
+            num_heads: self.num_heads,
+            num_layers: self.num_layers,
+            dim_feedforward: self.dim * 4,
+            causal: self.causal,
+            norm_first: true,
+            bias_ff: false,
+            bias_attn: false,
+            layer_scale: None,
+            context: self.context,
+            max_period: self.max_period as usize,
+            use_conv_block: false,
+            use_conv_bias: true,
+            cross_attention: None,
+            gating: Some(candle_nn::Activation::Silu),
+            norm: moshi::NormType::RmsNorm,
+            positional_embedding: moshi::transformer::PositionalEmbedding::Rope,
+            conv_layout: false,
+            conv_kernel_size: 3,
+            kv_repeat: 1,
+            max_seq_len: 4096 * 4,
+            shared_cross_attn: false,
+        };
+        moshi::lm::Config {
+            transformer: lm_cfg,
+            depformer: None,
+            audio_vocab_size: self.card + 1,
+            text_in_vocab_size: self.text_card + 1,
+            text_out_vocab_size: self.text_card,
+            audio_codebooks: self.n_q,
+            conditioners: Default::default(),
+            extra_heads: None,
+        }
+    }
+}
